dom/html/test/test_imports_nested.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=877072
-->
<head>
  <title>Test for Bug 877072</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=877072">Mozilla Bug 877072</a>

  <script type="text/javascript">
    SimpleTest.waitForExplicitFinish();
    var counter = 0;
    var fcounter = 0;
    var order = [];
    function loaded() {
      counter++;
    }
    function failed() {
      fcounter++;
    }
  </script>

  <link rel="import" href="imports/file_importA1.html" id="A1" onload="loaded()" onerror="failed()"></link>
  <link rel="import" href="imports/file_importB1.html" id="B1" onload="loaded()" onerror="failed()"></link>
  <link rel="import" href="imports/file_importB2.html" id="B2_2" onload="loaded()" onerror="failed()"></link>

  <script type="text/javascript">
    is(counter, 5, "Imports are loaded");
    is(fcounter, 0, "No error in imports");
    var expected = ["A2", "A1", "B2", "B1"];
    for (i in expected)
      is(order[i], expected[i], "import " + i + " should be " + expected[i]);
    SimpleTest.finish();
  </script>
</body>
</html>
